Move more methods to activity.cpp
