+++ src/activity.hpp
@@ -22,4 +22,15 @@ void continuous_action_eating();
 void continuous_action_eating_finish();
 void continuous_action_others();
 
+void select_random_fish();
+void get_fish();
+int search_material_spot();
+void spot_fishing();
+void spot_material();
+void spot_digging();
+void spot_mining_or_wall();
+turn_result_t do_dig_after_sp_check();
+void matdelmain(int = 0, int = 0);
+void matgetmain(int = 0, int = 0, int = 0);
+
 } // namespace elona
