+++ src/variables.hpp
@@ -759,7 +759,6 @@ int label_2073();
 int can_do_ranged_attack();
 int label_2083();
 int ask_direction();
-int search_material_spot();
 int decode_book();
 int read_normal_book();
 int label_2167();
@@ -1094,13 +1093,6 @@ void label_2150();
 void label_2151();
 void do_rest();
 void label_2153();
-void select_random_fish();
-void get_fish();
-void spot_fishing();
-void spot_material();
-void spot_digging();
-void spot_mining_or_wall();
-void continuous_action_eat();
 void label_2161();
 void show_eating_message();
 void heal_both_rider_and_mount();
@@ -1119,7 +1111,6 @@ void do_physical_attack();
 void proc_weapon_enchantments();
 void discover_trap();
 void discover_hidden_path();
-turn_result_t do_dig_after_sp_check();
 turn_result_t do_plant();
 void grow_plant(int);
 void try_to_grow_plant(int = 0);
@@ -1170,9 +1161,7 @@ void map_converttile();
 void map_createroomdoor();
 void map_initcustom(const std::string&);
 void map_initialize();
-void matdelmain(int = 0, int = 0);
 turn_result_t do_use_magic();
-void matgetmain(int = 0, int = 0, int = 0);
 void modrank(int = 0, int = 0, int = 0);
 void msg_clear();
 void msg_halt();
